refactor: libdebugの消去、使用関数の移動、使用箇所の変更
