+++ Makefile
@@ -16,7 +16,7 @@
 # ***********************************
 
 NAME		= minishell
-includes	= ./includes ./libft/libft ./libft/libex ./libft/libhash ./libft/libdebug
+includes	= ./includes ./libft/libft ./libft/libex ./libft/libhash
 src_dir		= srcs
 obj_dir		= objs
 obj			= $(src:%.c=$(src_dir)/%.o)
@@ -24,7 +24,7 @@ obj			= $(src:%.c=$(src_dir)/%.o)
 # ***********************************
 
 CC 			= gcc
-LIBS		= -L./libft/libft -L./libft/libex -L./libft/libhash -L./libft/libdebug -lft -lex -lhash -lreadline
+LIBS		= -L./libft/libft -L./libft/libex -L./libft/libhash -lft -lex -lhash -lreadline
 CFLAGS		= -Wall -Wextra -Werror -O3 -g $(includes:%=-I%)
 
 # ***********************************
@@ -63,11 +63,6 @@ libex		= $(libex_dir)/libex.a
 libhash_dir	= $(lib_dir)/libhash
 libhash		= $(libhash_dir)/libhash.a
 
-# ****************
-
-libdebug_dir	= $(lib_dir)/libdebug
-libdebug		= $(libdebug_dir)/libdebug.a
-
 # ***********************************
 
 all			: $(NAME)
@@ -92,7 +87,6 @@ init		:
 	zsh header.sh $(libft_dir) $(libft_dir)/libft.h $(libft_dir)/Makefile
 	zsh header.sh $(libex_dir) $(libex_dir)/libex.h $(libex_dir)/Makefile
 	zsh header.sh $(libhash_dir) $(libhash_dir)/libhash.h $(libhash_dir)/Makefile
-	zsh header.sh $(libdebug_dir) $(libdebug_dir)/libdebug.h $(libdebug_dir)/Makefile
 
 get_module	:
 	git submodule update --init
@@ -106,11 +100,11 @@ test_unit	: get_module
 test_issue	: get_module
 	bash ./tests/all-test.sh ./tests/issue $(TARGET)
 
-leak		: $(obj) $(lib) $(libdebug)
-	$(CC) $(CFLAGS) $(obj) $(lib) $(libdebug) $(sharedlib) -o $(NAME) -lreadline
+leak		: $(obj) $(lib)
+	$(CC) $(CFLAGS) $(obj) $(LIBS) $(sharedlib) -o $(NAME)
 
 sani-debug	: fclean lib_sani-debug
-	$(MAKE) CFLAGS="$(CFLAGS) -fsanitize=address" lib="$(lib) $(libdebug)"
+	$(MAKE) CFLAGS="$(CFLAGS) -fsanitize=address" $(lib)
 	$(MAKE) clean
 
 norm		:
@@ -119,13 +113,13 @@ norm		:
 	@printf "\e[32m%s\n\e[m" "Norm OK!"
 
 tests/%/main.c: $(lib) FORCE
-	$(CC) $(CFLAGS) $@ $(filter-out srcs/./main.c,$(src:%.c=$(src_dir)/%.c)) -o $(subst main.c,a.out,$@) $(LIBS) -ldebug
+	$(CC) $(CFLAGS) $@ $(filter-out srcs/./main.c,$(src:%.c=$(src_dir)/%.c)) -o $(subst main.c,a.out,$@) $(LIBS)
 
 tests/%/main.c+leak: $(lib) $(sharedlib)
-	$(CC) $(CFLAGS) $(subst +leak,,$@) $(filter-out srcs/./main.c,$(src:%.c=$(src_dir)/%.c)) $(sharedlib) -o $(subst main.c+leak,a.out,$@) $(LIBS) -ldebug
+	$(CC) $(CFLAGS) $(subst +leak,,$@) $(filter-out srcs/./main.c,$(src:%.c=$(src_dir)/%.c)) $(sharedlib) -o $(subst main.c+leak,a.out,$@) $(LIBS)
 
-tests/%/main.c+sani: $(lib) $(libdebug)
-	$(CC) $(CFLAGS) -fsanitize=address $(subst +sani,,$@) $(filter-out srcs/./main.c,$(src:%.c=$(src_dir)/%.c)) -o $(subst main.c+sani,a.out,$@) $(LIBS) -ldebug
+tests/%/main.c+sani: $(lib)
+	$(CC) $(CFLAGS) -fsanitize=address $(subst +sani,,$@) $(filter-out srcs/./main.c,$(src:%.c=$(src_dir)/%.c)) -o $(subst main.c+sani,a.out,$@) $(LIBS)
 
 prepush		: norm test
 
@@ -140,27 +134,21 @@ $(libex): $(libex_dir)/*.c
 $(libhash): $(libhash_dir)/*.c
 	$(MAKE) -C $(libhash_dir)
 
-$(libdebug): $(libdebug_dir)/*.c
-	$(MAKE) -C $(libdebug_dir)
-
 lib_make	:$(libft) $(libex) $(libhash)
 
 lib_clean	:
 	$(MAKE) clean -C $(libft_dir)
 	$(MAKE) clean -C $(libex_dir)
 	$(MAKE) clean -C $(libhash_dir)
-	$(MAKE) clean -C $(libdebug_dir)
 
 lib_fclean	:
 	$(MAKE) fclean -C $(libft_dir)
 	$(MAKE) fclean -C $(libex_dir)
 	$(MAKE) fclean -C $(libhash_dir)
-	$(MAKE) fclean -C $(libdebug_dir)
 
 lib_sani-debug	:
 	$(MAKE) -C $(libft_dir)
 	$(MAKE) sani-debug -C $(libex_dir)
 	$(MAKE) sani-debug -C $(libhash_dir)
-	$(MAKE) sani-debug -C $(libdebug_dir)
 
 FORCE:
