--- libft/libdebug/Makefile
@@ -1,36 +1,0 @@
-# ***********************************
-
-NAME	= libdebug.a
-CC		= gcc
-CFLAGS	= -Wall -Wextra -Werror -O3 -g
-obj		= $(src:%.c=%.o)
-
-.PHONY: all clean fclean re debug sani-debug
-
-# ***********************************
-
-src =\
-	./debug_array.c \
-	./debug_array_sort.c \
-	./debug_hash_display_table.c \
-
-# ***********************************
-
-all: $(NAME)
-
-$(NAME): $(obj)
-	$(AR) -rc $(NAME) $(obj)
-	ranlib $(NAME)
-
-clean:
-	$(RM) $(obj)
-
-fclean: clean
-	$(RM) $(NAME)
-
-re: fclean all
-
-# ***********************************
-
-sani-debug: fclean
-	$(MAKE) CFLAGS="$(CFLAGS) -fsanitize=address"
