+++ libft/libex/libex.h
@@ -33,5 +33,10 @@ int		ft_strcmp(const char *s1, const char *s2);
 char	*ft_xstrdup(const char *s1);
 bool	is_space_string(char *s);
 size_t	spacelen(char *s);
+int		arraylen(char **array);
+void	print_array(char **array);
+char	*random_string(size_t size);
+char	**str_arraydup(char **array);
+int		debug_arraycmp(char **aa, char **bb);
 
 #endif
