+++ libft/libex/print_array.c
@@ -1,0 +1,19 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   display_array.c                                    :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: kohkubo <kohkubo@student.42tokyo.jp>       +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/08/19 11:41:53 by kohkubo           #+#    #+#             */
+/*   Updated: 2021/08/19 11:42:13 by kohkubo          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "libex.h"
+
+void	print_array(char **array)
+{
+	while (*array)
+		printf("%s\n", *array++);
+}
